Update test and accept invite controller
